Resolve patch failures with CM
If you use these source patches and this (or later) commit causes you grief, please open an issue. If I have
to I can give each major flavor a branch. My preference and goal is to keep the universal patches working,
but all of the Android world is busy replicating work, fighting over licensing and consolidating power. So we
have a ton of competing aims and differences between firmware that only exist in order to create those differences.